requirements.txt
Flask==2.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.7
Werkzeug==2.0.3